requirements.txt
discord.py>=2.3
wavelink==3.3.0
python-dotenv>=1.0
